+++ src/ragger/firmware/structs.py
@@ -21,6 +21,7 @@ class Firmware(IntEnum):
     NANOSP = auto()
     NANOX = auto()
     STAX = auto()
+    FLEX = auto()
 
     @property
     def device(self) -> str:
@@ -43,19 +44,3 @@ class Firmware(IntEnum):
         States if the firmware's name starts with 'nano' or not.
         """
         return self.name.startswith("nano")
-
-    @property
-    def has_bagl(self):
-        """
-        States if the firmware uses BAGL or not.
-        Currently, this is a proxy to :attr:`.Firmware.is_nano`
-        """
-        return self.is_nano
-
-    @property
-    def has_nbgl(self):
-        """
-        States if the firmware uses NBGL or not.
-        Currently only Stax device uses NBGL.
-        """
-        return self.name.startswith("stax")