+++ src/ragger/gui/interface.py
@@ -77,7 +77,7 @@ class RaggerMainWindow(QMainWindow):
         self._screenshot.setObjectName("screenshot")
         self._screenshot.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
         self._screenshot.setAlignment(Qt.AlignVCenter)
-        dict_margin = {"nanos": 75, "nanox": 90, "nanosp": 55, "stax": 65}
+        dict_margin = {"nanos": 75, "nanox": 90, "nanosp": 55, "stax": 65, "flex": 65}
         margin = dict_margin[self._device]
         self._screenshot.setStyleSheet(f"QLabel {{margin-left: {margin}px;}}")
 
@@ -194,7 +194,7 @@ class RaggerMainWindow(QMainWindow):
         self._init_screenshot()
         self._init_validation_buttons()
 
-        if self._device == "stax":
+        if self._device in ["stax", "flex"]:
             self._devicebody.raise_()
             self._touch.raise_()