+++ src/ragger/navigator/touch_navigator.py
@@ -18,16 +18,16 @@ from typing import Callable, Dict
 
 from ragger.backend import BackendInterface
 from ragger.firmware import Firmware
-from ragger.firmware.stax import FullScreen
+from ragger.firmware.touch import FullScreen
 from .navigator import Navigator
 from .instruction import BaseNavInsID, NavInsID
 
 
-class StaxNavigator(Navigator):
+class TouchNavigator(Navigator):
 
     def __init__(self, backend: BackendInterface, firmware: Firmware, golden_run: bool = False):
-        if firmware != Firmware.STAX:
-            raise ValueError(f"'{self.__class__.__name__}' only works on Stax")
+        if firmware not in [Firmware.STAX, Firmware.FLEX]:
+            raise ValueError(f"'{self.__class__.__name__}' only works on Stax or Flex")
         screen = FullScreen(backend, firmware)
         callbacks: Dict[BaseNavInsID, Callable] = {
             NavInsID.WAIT: sleep,